requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.11
dukascopy-python>=1.0.0
requests>=2.31
python-telegram-bot>=20.0
loguru>=0.7
python-dotenv>=1.0
